refactor: moved some utilities in dedicated utils file
